pyproject.toml
[tool.poetry]
name = "eventsourcing-fastapi-graphql"
version = "0.1.0"
description = ""
authors = ["Martin R <apurplethinker@gmail.com>"]
license = "BSD 3"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
eventsourcing = "^9.2.22"
fastapi = "^0.111.0"
strawberry-graphql = "^0.234.0"
uvicorn = {version = "^0.27.0.post1", optional = true}
pydantic-settings = "^2.3.1"
psycopg2-binary = "^2.9.9"

[tool.poetry.extras]
production = ["uvicorn"]

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
ruff = "^0.4.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
